Replace distutils use in libcxx
Distutils is deprecated and slated for removal in python 3.12.
Debian-based distros have stopped packaging it in the stdlib by default
as well. libcxx merge_archives.py was using it to find paths to
executables. The best replacement is `shutil.which`, which also
correctly honors the `PATHEXT` environment variable on Windows instead
of append `.exe` to all of the executables. Note that `shutil.which` is
not available in python 2.7.