[lit] Add redirecting stdout into stderr to lit
lit tests only seem to report stderr by default. Sometimes you want to
see the stdout for a given test to see things like remarks when
attempting to debug a failure that only reproduces once in a blue moon
on CI when no one is looking. In those cases, you want to redirect
stdout to stderr so that lit will show you everything in the failure.